fix: Enable compose repos during build
In recent builds of achillobator we noticed that installing
systemd-container downgrades the systemd package already built into the
image.  This is because the centos-bootc image we're basing on has a
newer systemd than what is on the mirror network.  I believe this is
caused by the centos-bootc image enabling the compose repos during its
build.

https://gitlab.com/redhat/centos-stream/containers/bootc/-/blob/c10s/cs.repo

To avoid this, we can enable the same repos used during the centos-bootc
build.  To avoid stomping on the default repos, we'll change the repo
IDs.  We'll also disable these repos at the end of the build.